+++ examples/use_rust/esp32c6_ble/src/main.rs
@@ -5,10 +5,7 @@ mod macros;
 mod keymap;
 mod vial;
 
-use crate::{
-    keymap::{COL, NUM_LAYER, ROW},
-    vial::{VIAL_KEYBOARD_DEF, VIAL_KEYBOARD_ID},
-};
+use crate::vial::{VIAL_KEYBOARD_DEF, VIAL_KEYBOARD_ID};
 use defmt::*;
 use esp_idf_svc::hal::{gpio::*, peripherals::Peripherals, task::block_on};
 use esp_println as _;