+++ rmk-macro/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- ### Changed
+
+- BREAKING: generate new public APIs
 - Add support for overriding chip initialization for nrf/rp/esp32
 - Fix broken link in error message