+++ rmk-macro/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-09-11
+
 ### Added
 
 - Add support for split keyboard(serial/BLE)